+++ R/xml.R
@@ -1,4 +1,18 @@
 # ____________----
+# HELPERS ---------------------------------------------
+
+emu_to_inch <- function(x) {
+  x / 914400
+}
+
+
+cm_to_inch <- function(x) {
+  x * .39370078740157
+}
+
+
+
+# ____________----
 # XML --------------------------------------------
 
 
@@ -13,6 +27,19 @@ xml_get_shapes <- function(x) {
 # x : node or nodeset
 xml_get_runs <- function(x) {
   xml_find_all(x, ".//a:p/a:r")
+}
+
+
+# filter shapes by text pattern
+# x : node or nodeset
+# returns: shape nodes
+xml_shapes_filter <- function(shapes, pattern = NULL) {
+  if (is.null(pattern)) {
+    return(shapes)
+  }
+  texts <- shapes |> xml_text()
+  ii <- stringr::str_detect(texts, stringr::fixed(pattern))
+  shapes[ii]
 }
 
 
@@ -66,6 +93,21 @@ xml_shape_remove <- function(shape) {
 }
 
 
+read_xfrm <- getFromNamespace("read_xfrm", ns = "officer")
+
+# get xfrm info fron shape node
+# xfrm into: http://officeopenxml.com/drwSp-size.php
+# TODO: placeholders to not always have xrfm
+xml_shape_get_frame <- function(shape) {
+  l <- read_xfrm(shape, "", NA) |> unlist()
+  p <- l[c("offx", "offy", "cx", "cy")] |>
+    as.numeric() |>
+    emu_to_inch() |>
+    setNames(c("left", "top", "width", "height")) |>
+    as.list()
+  p
+}
+
 
 # ____________----
 # PPTX --------------------------------------------
@@ -79,6 +121,41 @@ pptx_slide_get <- function(x, slide_idx) {
 
 
 # get all shapes on slide
-pptx_shapes_on_slide <- function(x, slide_idx) {
-  pptx_slide_get(x, slide_idx) |> xml_get_shapes()
+pptx_shapes_on_slide <- function(x, slide_idx, pattern = NULL) {
+  shapes <- pptx_slide_get(x, slide_idx) |> xml_get_shapes()
+  shapes |> xml_shapes_filter(pattern)
+}
+
+
+#' Hide und unhide shapes that match pattern
+#' @param x `[rpptx]`\cr An [officer] object. See [read_pptx()].
+#' @param pattern `[character]`\cr Vector with patterns to find shapes to hide/unhide. Regex is not interpreted.
+#' @param slide_idx `[numeric]`\cr Index of slides to process. If `NULL` (default), all slides
+#' @export
+#' @rdname shapes-hide-unhide
+pptx_shapes_hide_temp <-function(x, pattern, slide_idx = NULL) {
+  .pptx_hide_unhide_shapes("hide", x, pattern, slide_idx)
+}
+
+
+#' @export
+#' @rdname shapes-hide-unhide
+pptx_shapes_unhide_temp <-function(x, pattern, slide_idx = NULL) {
+  .pptx_hide_unhide_shapes("unhide", x, pattern, slide_idx)
+}
+
+
+.pptx_hide_unhide_shapes <- function(action = "hide", x, pattern, slide_idx = NULL, ...) {
+  assert_class(x, "rpptx")
+  action <- match.arg(action, c("hide", "unhide"))
+  .fun <- switch(action, hide = xml_shape_hide, unhide = xml_shape_unhide)
+  slide_idx <- slide_idx %||% seq_along(x)
+  for (s_idx in slide_idx) {
+    x$cursor <- s_idx  # necessary for officer functions
+    for (i in seq_along(pattern)) {
+      shapes <- x |> pptx_shapes_on_slide(slide_idx = s_idx, pattern = pattern[i])
+      .fun(shapes)
+    }
+  }
+  x
 }