+++ inst/ext/examples/example-text-replace.R
@@ -1,11 +1,11 @@
 # Some example file
 
-file_in <- example_file("example_01")
+file_in <- example_file("text_replace")
 file_out <- tempfile(fileext = ".pptx")
 
 # file_open(file_in)
 
-x <- read_pptx(file_in)
+x <- pptx_read(file_in)
 x <- pptx_text_replace(x, "@" = ">>>", "{1}" = ":)")
 print(x, target = file_out)