+++ NEWS.md
@@ -1,5 +1,10 @@
 # officer.pptx 0.1.0
 
-* `example_file` and `test_file` to find example and test files.
+* WIP: `pptx_slide_draw`: draw wireframe-like image of elements (shapes, placeholders, images) on a slide.
+* `pptx_image_insert_at_shape_temp`, `pptx_shapes_hide_temp`, and `pptx_shapes_unhide_temp` as drafts
+used in production. Also several helper functions.
+* new tests
+* `frame` class to represent a bounding box (used for shapes etc.) and associated functions `frame_*`
+* `example_file`, `test_file`, and `test_image` to find example and test files.
 * added several xml and pptx helper functions incl. tests
 * `pptx_text_replace`: replace text patterns in powerpoint slides. Formatting is respected.